src/main/java/com/travel/api/service/ProductService.java
package com.travel.api.service;

import java.util.List;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.travel.api.repository.ProductRepository;
import com.travel.api.vo.Product;

@Service
public class ProductService {
    private ProductRepository productRepository;

    @Autowired
    public ProductService(ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> getProducts() {
        return productRepository.findAll();
    }

    public void save(Product product) {
        productRepository.save(product);
    }

    public Product findById(long id) {
        Optional<Product> byId = productRepository.findById(id);
        return byId.orElse(null);
    }
}
